+++ utoipa-gen/src/component/schema.rs
@@ -9,10 +9,7 @@ use syn::{
 
 use crate::{
     as_tokens_or_diagnostics,
-    component::features::{
-        attributes::{Rename, Title, ValueType},
-        validation::Pattern,
-    },
+    component::features::attributes::{Rename, Title, ValueType},
     doc_comment::CommentAttributes,
     parse_utils::LitBoolOrExprPath,
     Array, AttributesExt, Diagnostics, OptionExt, ToTokensDiagnostics,
@@ -744,20 +741,31 @@ impl UnnamedStructSchema {
                     ));
                 }
             }
-            let pattern = if let Some(pattern) =
-                pop_feature!(features => Feature::Pattern(_) as Option<Pattern>)
-            {
-                // Pattern Attribute is only allowed for unnamed structs with single field
-                if fields_len > 1 {
+
+            if fields_len > 1 {
+                // The struct has multiple unnamed fields.
+                // Validation related features are only allowed on unnamed structs with a single field (i.e., newtype pattern).
+                if let Some((name, span)) = features.iter().find_map(|feature| match feature {
+                    Feature::MultipleOf(attr) => Some(("multiple_of", attr.span())),
+                    Feature::Maximum(attr) => Some(("maximum", attr.span())),
+                    Feature::Minimum(attr) => Some(("minimum", attr.span())),
+                    Feature::ExclusiveMaximum(attr) => Some(("exclusive_maximum", attr.span())),
+                    Feature::ExclusiveMinimum(attr) => Some(("exclusive_minimum", attr.span())),
+                    Feature::MinLength(attr) => Some(("min_length", attr.span())),
+                    Feature::MaxLength(attr) => Some(("max_length", attr.span())),
+                    Feature::Pattern(attr) => Some(("pattern", attr.span())),
+                    Feature::MaxItems(attr) => Some(("max_items", attr.span())),
+                    Feature::MinItems(attr) => Some(("min_items", attr.span())),
+                    _ => None,
+                }) {
                     return Err(Diagnostics::with_span(
-                        pattern.span(),
-                        "Pattern attribute is not allowed for unnamed structs with multiple fields",
+                        span,
+                        format!(
+                            "{name} attribute is not allowed for unnamed structs with multiple fields",
+                        ),
                     ));
                 }
-                Some(pattern.to_token_stream())
-            } else {
-                None
-            };
+            }
 
             let comments = CommentAttributes::from_attributes(root.attributes);
             let description = description
@@ -780,11 +788,7 @@ impl UnnamedStructSchema {
             })?;
 
             tokens.extend(schema.to_token_stream());
-            if let Some(pattern) = pattern {
-                tokens.extend(quote! {
-                    .pattern(Some(#pattern))
-                });
-            }
+
             schema_references = std::mem::take(&mut schema.schema_references);
         } else {
             // Struct that has multiple unnamed fields is serialized to array by default with serde.